refactor(callbacks): migrate to student nomenclature
