+++ professor-virtual/professor_virtual/shared_libraries/__init__.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 """ includes all shared libraries for the agent."""
 from .callbacks.rate_limit_callback import rate_limit_callback
-from .callbacks.validate_customer_id import validate_customer_id
+from .callbacks.validate_customer_id import validate_student_id
 from .callbacks.lowercase_value import lowercase_value
 from .callbacks.before_tool import before_tool
 from .callbacks.after_tool import after_tool
@@ -22,7 +22,7 @@ from .callbacks.before_agent import before_agent
 
 __all__ = [
     "rate_limit_callback",
-    "validate_customer_id",
+    "validate_student_id",
     "lowercase_value",
     "before_tool",
     "after_tool",
