+++ professor-virtual/professor_virtual/shared_libraries/callbacks/__init__.py
@@ -1,5 +1,5 @@
 from .rate_limit_callback import rate_limit_callback
-from .validate_customer_id import validate_customer_id
+from .validate_customer_id import validate_student_id
 from .lowercase_value import lowercase_value
 from .before_tool import before_tool
 from .after_tool import after_tool
@@ -7,7 +7,7 @@ from .before_agent import before_agent
 
 __all__ = [
     "rate_limit_callback",
-    "validate_customer_id",
+    "validate_student_id",
     "lowercase_value",
     "before_tool",
     "after_tool",
