+++ professor-virtual/professor_virtual/shared_libraries/callbacks/before_tool/before_tool_callback.py
@@ -3,7 +3,7 @@ from typing import Any, Dict
 from google.adk.tools import BaseTool
 from google.adk.agents.callback_context import CallbackContext
 from ..lowercase_value import lowercase_value
-from ..validate_customer_id import validate_customer_id
+from ..validate_customer_id import validate_student_id
 
 logger = logging.getLogger(__name__)
 
@@ -15,11 +15,11 @@ def before_tool(
     # i make sure all values that the agent is sending to tools are lowercase
     lowercase_value(args)
 
-    # Several tools require customer_id as input. We don't want to rely
-    # solely on the model picking the right customer id. We validate it.
-    # Alternative: tools can fetch the customer_id from the state directly.
-    if 'customer_id' in args:
-        valid, err = validate_customer_id(args['customer_id'], tool_context.state)
+    # Several tools require student_id as input. We don't want to rely
+    # solely on the model picking the right student id. We validate it.
+    # Alternative: tools can fetch the student_id from the state directly.
+    if 'student_id' in args:
+        valid, err = validate_student_id(args['student_id'], tool_context.state)
         if not valid:
             return err
